se agrego la validacion para la eliminacion de area

+++ src/Controller/AreaController.php
@@ -122,6 +122,14 @@ class AreaController extends AbstractController
      */
     public function delete(Request $request, Area $area, EntityManagerInterface $entityManager): Response
     {
+        if(!empty($area->getTipoNormaReparticions()->toArray())){
+            $this->addFlash(
+                'errorDeleteArea',
+                "No puede eliminar esta repartición ya que está vinculada a un tipo de norma."
+            );
+            return $this->redirectToRoute('area_index',[],Response::HTTP_SEE_OTHER);
+        }
+
         if ($this->isCsrfTokenValid('delete'.$area->getId(), $request->request->get('_token'))) {
             $entityManager->remove($area);
             $entityManager->flush();
